remove SupportConfigEmailPropertyService.js using emailutility instead
