+++ src/foam/nanos/notification/email/EmailPropertyService.js
@@ -32,11 +32,6 @@ foam.INTERFACE({
           name: 'templateArgs',
           type: 'Map',
           documentation: 'Template arguments'
-        },
-        {
-          name: 'theme',
-          type: 'foam.nanos.theme.Theme',
-          documentation: 'theme'
         }
       ],
     }
