+++ src/foam/util/Emails/EmailsUtility.java
@@ -3,6 +3,7 @@ package foam.util.Emails;
 import foam.core.X;
 import foam.dao.DAO;
 import foam.nanos.app.AppConfig;
+import foam.nanos.app.EmailConfig;
 import foam.nanos.auth.Subject;
 import foam.nanos.auth.User;
 import foam.nanos.logger.Logger;
@@ -59,6 +60,18 @@ public class EmailsUtility {
     }
 
     SupportConfig supportConfig = theme.getSupportConfig();
+
+    // Set ReplyTo, From, DisplayName in emailMessage 
+    EmailConfig emailConfig = supportConfig.getEmailConfig();
+    // REPLY TO:
+    emailMessage.setReplyTo(emailConfig.getReplyTo());
+
+    // DISPLAY NAME:
+    emailMessage.setDisplayName(emailConfig.getDisplayName());
+
+    // FROM:
+    emailMessage.setFrom(emailConfig.getFrom());
+
     // Add template name to templateArgs, to avoid extra parameter passing
     if ( ! SafetyUtil.isEmpty(templateName) ) {
       if ( templateArgs != null ) {
@@ -71,13 +84,13 @@ public class EmailsUtility {
       templateArgs.put("supportEmail", (supportConfig.getSupportEmail()));
 
       // personal support user
-      User psUser = theme.findPersonalSupportUser(x);
+      User psUser = supportConfig.findPersonalSupportUser(x);
       templateArgs.put("personalSupportPhone", psUser == null ? "" : psUser.getPhoneNumber());
       templateArgs.put("personalSupportEmail", psUser == null ? "" : psUser.getEmail());
       templateArgs.put("personalSupportFirstName", psUser == null ? "" : psUser.getFirstName());
       templateArgs.put("personalSupportFullName", psUser == null ? "" : psUser.getLegalName());
 
-      foam.nanos.auth.Address address = theme.getSupportAddress();
+      foam.nanos.auth.Address address = supportConfig.getSupportAddress();
       templateArgs.put("supportAddress", address == null ? "" : address.toSummary());
       templateArgs.put("appName", (theme.getAppName()));
       templateArgs.put("logo", (appConfig.getUrl() + "/" + theme.getLogo()));
@@ -88,7 +101,7 @@ public class EmailsUtility {
     // SERVICE CALL: to fill in email properties.
     EmailPropertyService cts = (EmailPropertyService) x.get("emailPropertyService");
     try {
-      cts.apply(x, group, emailMessage, templateArgs, theme);
+      cts.apply(x, group, emailMessage, templateArgs);
     } catch (Exception e) {
       logger.error(e);
       return;
